20/01/2020 10:46

+++ src/main/java/com/lanhuongcosmetic/dao/IProductDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface IProductDAO extends GenericDAO<ProductModel> {
     List<ProductModel> findByCategory(int category_id);
-    Long save(ProductModel productModel);
+    int save(ProductModel productModel);
     void update(ProductModel productModel);
     void delete(int product_id);
     ProductModel findOne(int product_id);
