+++ src/main/java/com/lanhuongcosmetic/service/IProductService.java
@@ -8,5 +8,5 @@ public interface IProductService {
     List<ProductModel> findByCategory(int category_id);
     ProductModel save(ProductModel productModel);
     ProductModel update(ProductModel productModel);
-    ProductModel findOne(int product_id);
+    ProductModel findOne(long product_id);
 }
